+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Spawner extraction machine
 - KubeJS plugin for generating new cooling coils
 - Fluid cooling recipes
+- Cooling tower
 
 ### Changed